video_processing/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
mediapipe==0.10.14
opencv-contrib-python==4.10.0.84
python-dotenv==1.0.1
openai==1.43.0
anthropic==0.37.1
deepgram-sdk==3.5.1
boto3==1.35.8
moviepy==1.0.3
Werkzeug==3.0.4
resend==2.4.0
gunicorn[gthread]==21.2.0